requirements.txt
ibm-watsonx-ai
scikit-learn
matplotlib
elasticsearch
streamlit
loguru
python-dotenv